Adding extra samples
Addin the ability to see GPU-freq and ext-bus freq. This is somewhat too
vendor-specific. Will revert this and make a more generic solution ASAP where
extra sample can be defined much more flexably. This however pushes the limit
of what is doable in a sampler based on a script (either internally or
externally run) as sh-loops and logic are quite slow, which will affect
temporal accuracy of the sampler.

Catehegory: development